Restructure: compile() is complete (#38)
* changed the data type of Program::ids

* apply new structure to compile()

* remove debug logs

* minor changes

* compile() is complete

+++ control/inc/errorControl.h
@@ -15,7 +15,7 @@ public:
     void printToConsole(QString message);
     void printErrorMsg(ResultState result);
     void printErrorMsg(QString errorMsg);
-    void printErrorMsgAtLine(ResultState result, qint16 lineNum);
+    void printErrorMsgAtLine(ResultState result, int lineNum);
 
 private:
     QString convertErrorMsg(ResultState result);
